requirements.txt
torch
numpy
einops
torchvision
tqdm
diffusers
accelerate
transformers
Pillow
wandb
bitnet
peewee
git+https://github.com/openai/CLIP.git